Add files via upload
This is the re-vamp of greenbanker.com. I started designing the wireframe and user interface a week ago. I completed just about 15% of the home page today. With critiques and good input from colleagues, I believe this website will work well for the companies goals and my own in correlation with Facebook Ads and other forms of advertising, will lower our bounce rate and therefore, influence more clients to list with Green Banker.